+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Updated to work in Elasticsearch v8.5.2
 - Replaced the arg_parser code with gen_class.ArgParser code.
 
+### Changed
+- main, run_program: Removed gen_libs.get_inst call.
+- Documentation update
+
 
 ## [0.3.6] - 2022-12-22
 - Updated to work in Python 3 too